feat: rename conversation

+++ app/javascript/components/Menu.tsx
@@ -10,19 +10,35 @@ interface MenuProps {
 
 const Menu: React.FC<MenuProps> = ({ onShowSignInModal, conversations, isMobile = false }) => {
   const [status, setStatus] = useState("initial")
+  const [title, setTitle] = useState("")
 
   const isCurrent = (conversation) => {
     return window.location.pathname === `/chats/${conversation.id}` || !!conversation.current
   }
 
+  const checkKeyPress = (e, conversationId) => {
+    if (e.key === "Enter") {
+      confirmAction(conversationId)
+    }
+  }
+
   const renderTitle = (conversation, i) => {
     switch (status) {
       case "initial":
-        return conversation.title
+        return title || conversation.title
       case "pendingDelete":
         return `删除「${conversation.title}」？`
+      case "pendingEdit":
+        return (
+          <input
+            autoFocus
+            value={title}
+            onChange={(e) => setTitle(e.target.value)}
+            onKeyDown={(e) => checkKeyPress(e, conversation.id)}
+          />
+        )
       default:
-        return ""
+        return conversation.title
     }
   }
 
@@ -33,9 +49,23 @@ const Menu: React.FC<MenuProps> = ({ onShowSignInModal, conversations, isMobile 
     window.location.href = "/chats/new"
   }
 
-  const deleteConversation = async (conversationId) => {
-    await axios.delete(`/v1/conversations/${conversationId}`)
-    window.location.href = "/chats/new"
+  const confirmAction = async (conversationId) => {
+    if (status === "pendingDelete") {
+      await axios.delete(`/v1/conversations/${conversationId}`)
+      window.location.href = "/chats/new"
+    }
+
+    if (status === "pendingEdit") {
+      await axios.put(`/v1/conversations/${conversationId}`, { title: title })
+      setStatus("initial")
+    }
+  }
+
+  const onClickEdit = (conversation) => {
+    setStatus("pendingEdit")
+    if (!title) {
+      setTitle(conversation.title)
+    }
   }
 
   return (
@@ -116,11 +146,12 @@ const Menu: React.FC<MenuProps> = ({ onShowSignInModal, conversations, isMobile 
                       <div className="truncate">{renderTitle(conversation, i)}</div>
                       <div className="relative w-8 z-10"></div>
                       <div className="absolute flex right-1 z-10 visible gap-2">
-                        {status === "pendingDelete" ? (
+                        {status !== "initial" ? (
                           <>
                             <button
+                              type="submit"
                               className="pt-px hover:text-zinc-900 dark:hover:text-white"
-                              onClick={() => deleteConversation(conversation.id)}
+                              onClick={() => confirmAction(conversation.id)}
                             >
                               <svg
                                 xmlns="http://www.w3.org/2000/svg"
@@ -150,25 +181,46 @@ const Menu: React.FC<MenuProps> = ({ onShowSignInModal, conversations, isMobile 
                             </button>
                           </>
                         ) : (
-                          <button
-                            className="pt-px hover:text-zinc-900 dark:hover:text-white"
-                            onClick={() => setStatus("pendingDelete")}
-                          >
-                            <svg
-                              xmlns="http://www.w3.org/2000/svg"
-                              fill="none"
-                              viewBox="0 0 24 24"
-                              strokeWidth="1.5"
-                              stroke="currentColor"
-                              className="w-4 h-4"
+                          <>
+                            <button
+                              className="pt-px hover:text-zinc-900 dark:hover:text-white"
+                              onClick={() => onClickEdit(conversation)}
                             >
-                              <path
-                                strokeLinecap="round"
-                                strokeLinejoin="round"
-                                d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"
-                              />
-                            </svg>
-                          </button>
+                              <svg
+                                xmlns="http://www.w3.org/2000/svg"
+                                fill="none"
+                                viewBox="0 0 24 24"
+                                strokeWidth="1.5"
+                                stroke="currentColor"
+                                className="w-4 h-4"
+                              >
+                                <path
+                                  strokeLinecap="round"
+                                  strokeLinejoin="round"
+                                  d="M16.862 4.487l1.687-1.688a1.875 1.875 0 112.652 2.652L10.582 16.07a4.5 4.5 0 01-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 011.13-1.897l8.932-8.931zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0115.75 21H5.25A2.25 2.25 0 013 18.75V8.25A2.25 2.25 0 015.25 6H10"
+                                />
+                              </svg>
+                            </button>
+                            <button
+                              className="pt-px hover:text-zinc-900 dark:hover:text-white"
+                              onClick={() => setStatus("pendingDelete")}
+                            >
+                              <svg
+                                xmlns="http://www.w3.org/2000/svg"
+                                fill="none"
+                                viewBox="0 0 24 24"
+                                strokeWidth="1.5"
+                                stroke="currentColor"
+                                className="w-4 h-4"
+                              >
+                                <path
+                                  strokeLinecap="round"
+                                  strokeLinejoin="round"
+                                  d="M14.74 9l-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 01-2.244 2.077H8.084a2.25 2.25 0 01-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 00-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 013.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 00-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 00-7.5 0"
+                                />
+                              </svg>
+                            </button>
+                          </>
                         )}
                       </div>
                     </a>
